refactor: nav links hooks -> components
